+++ lib/cassandra/protocol.rb
@@ -84,8 +84,6 @@ require 'cassandra/protocol/requests/void_query_request'
 require 'cassandra/protocol/requests/prepare_request'
 require 'cassandra/protocol/requests/execute_request'
 require 'cassandra/protocol/cql_protocol_handler'
-require 'cassandra/protocol/timestamp/time_based_generator'
-require 'cassandra/protocol/timestamp/jruby_time_based_generator'
 require 'cassandra/protocol/v1'
 require 'cassandra/protocol/v3'
 require 'cassandra/protocol/v4'
